requirements.txt
numpy==1.22.0
opencv-python==4.5.5.62
Pillow==7.0.0
PyYAML==6.0
scipy==1.10.1
tensorboard==2.14.0
tqdm==4.66.1
imageio
plyfile
piq
dearpygui
lpips
pytorch_msssim
matplotlib
scikit-image
openmesh 
git+https://github.com/Po-Hsun-Su/pytorch-ssim.git